Also capture build operations when FS watching is unsupported
